Added amount-tipped message
This uses a regular expression (re.compile("D?(\d+) ?(?:D|doge)?", re.IGNORECASE)) to look for the amount tipped when the bot itself is tipped. If it finds an amount, it will provide a number of hours that the bot will run because of the tip (amount tipped / average_tip). If it doesn't find an amount, it replies with the previous message.